# docs/swagger.yaml
definitions:
  internal_controllers_webhook.GenericResponse:
    properties:
      message:
        description: Message provides a brief status message for the operation.
        type: string
    type: object
...
  internal_controllers_webhook.SubscriptionView:
    properties:
      createdAt:
        description: CreatedAt is when the subscription was created.
        type: string
      description:
        description: Description is the optional description from the webhook trigger.
        type: string
      vehicleTokenId:
        description: VehicleTokenID is the on-chain token ID of the vehicle.
        type: string
      webhookId:
        description: webhookID is the identifier of the webhook trigger.
        type: string
    type: object
  internal_controllers_webhook.UpdateWebhookRequest:
    properties:
      condition:
        description: Condition updates the CEL expression used to decide when to fire.
        type: string
      coolDownPeriod:
        description: CoolDownPeriod updates the minimum number of seconds between
          firings.
        type: integer
      description:
        description: Description updates the optional human-friendly explanation of
          the webhook.
        type: string
      displayName:
        description: DisplayName updates the user-friendly unique name per developer
          license.
        type: string
      metricName:
        description: MetricName updates the signal/event name used by the webhook.
        type: string
      status:
        description: Status updates the current state of the webhook (e.g. "enabled"
          or "Disabled").
        type: string
      targetURL:
        description: TargetURL updates the HTTPS endpoint that will receive callbacks.
        type: string
    type: object
  internal_controllers_webhook.UpdateWebhookResponse:
    properties:
      id:
        description: ID is the unique identifier of the updated webhook.
        type: string
      message:
        description: Message provides a brief status message for the operation.
        type: string
    type: object
...
info:
  contact: {}
  title: Vehicle Triggers API
  version: "1.0"
paths:
...
  /v1/webhooks/{webhookId}:
    delete:
      description: Deletes a webhook by its ID.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Webhook deleted successfully
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "404":
          description: Webhook not found
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete a webhook
      tags:
      - Webhooks
    get:
      description: Returns every vehicleTokenId currently subscribed.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all vehicles subscribed to a webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Updates the configuration of a webhook by its ID. The failure count
        is reset to 0 when updating a webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      - description: Webhook configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controllers_webhook.UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook updated successfully
          schema:
            $ref: '#/definitions/internal_controllers_webhook.UpdateWebhookResponse'
        "400":
          description: Invalid request payload
        "404":
          description: Webhook not found
        "500":
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update a webhook
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/subscribe/{vehicleTokenId}:
    post:
      consumes:
      - application/json
      description: Associates a vehicle with a specific event webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      - description: Vehicle Token ID
        in: path
        name: vehicleTokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Vehicle assigned
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign a vehicle to a webhook
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/subscribe/all:
    post:
      description: Subscribes every vehicle shared with this developer to the webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Count of subscribed vehicles
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe all shared vehicles
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/subscribe/list:
    post:
      consumes:
      - application/json
      description: Subscribes each vehicleTokenId to the webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Count of subscribed vehicles
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign multiple vehicles to a webhook from a list
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/unsubscribe/{vehicleTokenId}:
    delete:
      description: Removes a vehicle's subscription.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      - description: Vehicle Token ID
        in: path
        name: vehicleTokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Vehicle removed
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe a vehicle from a webhook
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/unsubscribe/all:
    delete:
      description: Removes every shared vehicle subscription for this webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Count of unsubscribed vehicles
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe all shared vehicles
      tags:
      - Webhooks
  /v1/webhooks/{webhookId}/unsubscribe/list:
    delete:
      consumes:
      - application/json
      description: Unsubscribes each vehicleTokenId from the webhook.
      parameters:
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Count of unsubscribed vehicles
          schema:
            $ref: '#/definitions/internal_controllers_webhook.GenericResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe multiple vehicles from a webhook using a list
      tags:
      - Webhooks
...
  /v1/webhooks/vehicles/{vehicleTokenId}:
    get:
      description: Retrieves all webhook subscriptions for a given vehicle.
      parameters:
      - description: Vehicle Token ID
        in: path
        name: vehicleTokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_controllers_webhook.SubscriptionView'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List subscriptions for a vehicle
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"